backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
mysql-connector-python==8.2.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0.post1
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
PyJWT==2.8.0
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0
anthropic==0.8.1
pytest==7.4.3
pytest-asyncio==0.21.1
